# tests: Adding t.Parallel() for tests in model package