--- a/model/histogram/float_histogram_test.go
+++ b/model/histogram/float_histogram_test.go
@@ -222,7 +222,11 @@ func TestFloatHistogramMul(t *testing.T) {
 	}
 
 	for _, c := range cases {
+		// capture the range variable to avoid issues with concurrency
+		c := c
 		t.Run(c.name, func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			require.Equal(t, c.expected, c.in.Mul(c.scale))
 			// Has it also happened in-place?
 			require.Equal(t, c.expected, c.in)
@@ -284,7 +288,11 @@ func TestFloatHistogramCopy(t *testing.T) {
 	}
 
 	for _, tcase := range cases {
+		// capture the range variable to avoid issues with concurrency
+		tcase := tcase
 		t.Run(tcase.name, func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			hCopy := tcase.orig.Copy()
 
 			// Modify a primitive value in the original histogram.
@@ -349,7 +357,11 @@ func TestFloatHistogramCopyTo(t *testing.T) {
 	}
 
 	for _, tcase := range cases {
+		// capture the range variable to avoid issues with concurrency
+		tcase := tcase
 		t.Run(tcase.name, func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			hCopy := &FloatHistogram{}
 			tcase.orig.CopyTo(hCopy)
 
@@ -362,6 +374,7 @@ func TestFloatHistogramCopyTo(t *testing.T) {
 }
 
 func assertDeepCopyFHSpans(t *testing.T, orig, hCopy, expected *FloatHistogram) {
+	t.Helper()
 	// Do an in-place expansion of an original spans slice.
 	orig.PositiveSpans = expandSpans(orig.PositiveSpans)
 	orig.PositiveSpans[len(orig.PositiveSpans)-1] = Span{1, 2}
@@ -544,7 +557,11 @@ func TestFloatHistogramDiv(t *testing.T) {
 	}
 
 	for _, c := range cases {
+		// capture the range variable to avoid issues with concurrency
+		c := c
 		t.Run(c.name, func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			require.Equal(t, c.expected, c.fh.Div(c.s))
 			// Has it also happened in-place?
 			require.Equal(t, c.expected, c.fh)
@@ -1274,7 +1291,11 @@ func TestFloatHistogramDetectReset(t *testing.T) {
 	}
 
 	for _, c := range cases {
+		// capture the range variable to avoid issues with concurrency
+		c := c
 		t.Run(c.name, func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			require.Equal(t, c.resetExpected, c.current.DetectReset(c.previous))
 		})
 	}
@@ -1631,7 +1652,11 @@ func TestFloatHistogramCompact(t *testing.T) {
 	}
 
 	for _, c := range cases {
+		// capture the range variable to avoid issues with concurrency
+		c := c
 		t.Run(c.name, func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			require.Equal(t, c.expected, c.in.Compact(c.maxEmptyBuckets))
 			// Compact has happened in-place, too.
 			require.Equal(t, c.expected, c.in)
@@ -2309,7 +2334,11 @@ func TestFloatHistogramAdd(t *testing.T) {
 	}
 
 	for _, c := range cases {
+		// capture the range variable to avoid issues with concurrency
+		c := c
 		t.Run(c.name, func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			testHistogramAdd(t, c.in1, c.in2, c.expected, c.expErrMsg)
 			testHistogramAdd(t, c.in2, c.in1, c.expected, c.expErrMsg)
 		})
@@ -2317,6 +2346,7 @@ func TestFloatHistogramAdd(t *testing.T) {
 }
 
 func testHistogramAdd(t *testing.T, a, b, expected *FloatHistogram, expErrMsg string) {
+	t.Helper()
 	var (
 		aCopy        = a.Copy()
 		bCopy        = b.Copy()
@@ -2502,7 +2532,11 @@ func TestFloatHistogramSub(t *testing.T) {
 	}
 
 	for _, c := range cases {
+		// capture the range variable to avoid issues with concurrency
+		c := c
 		t.Run(c.name, func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			testFloatHistogramSub(t, c.in1, c.in2, c.expected, c.expErrMsg)
 
 			var expectedNegative *FloatHistogram
@@ -2515,6 +2549,7 @@ func TestFloatHistogramSub(t *testing.T) {
 }
 
 func testFloatHistogramSub(t *testing.T, a, b, expected *FloatHistogram, expErrMsg string) {
+	t.Helper()
 	var (
 		aCopy        = a.Copy()
 		bCopy        = b.Copy()
@@ -2627,7 +2662,11 @@ func TestFloatHistogramCopyToSchema(t *testing.T) {
 	}
 
 	for _, c := range cases {
+		// capture the range variable to avoid issues with concurrency
+		c := c
 		t.Run(c.name, func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			inCopy := c.in.Copy()
 			require.Equal(t, c.expected, c.in.CopyToSchema(c.targetSchema))
 			// Check that the receiver histogram was not mutated:
@@ -2637,6 +2676,8 @@ func TestFloatHistogramCopyToSchema(t *testing.T) {
 }
 
 func TestReverseFloatBucketIterator(t *testing.T) {
+	// mark this test case as being run in parallel
+	t.Parallel()
 	h := &FloatHistogram{
 		Count:         405,
 		ZeroCount:     102,
@@ -2880,7 +2921,11 @@ func TestAllFloatBucketIterator(t *testing.T) {
 	}
 
 	for i, c := range cases {
+		// capture the range variable to avoid issues with concurrency
+		c := c
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			var expBuckets, actBuckets []Bucket[float64]
 
 			if c.includeNeg {
@@ -3106,7 +3151,11 @@ func TestAllReverseFloatBucketIterator(t *testing.T) {
 	}
 
 	for i, c := range cases {
+		// capture the range variable to avoid issues with concurrency
+		c := c
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			var expBuckets, actBuckets []Bucket[float64]
 
 			if c.includePos {
@@ -3150,6 +3199,8 @@ func TestAllReverseFloatBucketIterator(t *testing.T) {
 }
 
 func TestFloatBucketIteratorTargetSchema(t *testing.T) {
+	// mark this test case as being run in parallel
+	t.Parallel()
 	h := FloatHistogram{
 		Count:  405,
 		Sum:    1008.4,
@@ -3244,7 +3295,11 @@ func TestFloatCustomBucketsIterators(t *testing.T) {
 	}
 
 	for i, c := range cases {
+		// capture the range variable to avoid issues with concurrency
+		c := c
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			{
 				it := c.h.AllBucketIterator()
 				for i, b := range c.expPositiveBuckets {
@@ -3303,6 +3358,8 @@ func TestFloatCustomBucketsIterators(t *testing.T) {
 // TestFloatHistogramEquals tests FloatHistogram with float-specific cases that
 // cannot be covered by TestHistogramEquals.
 func TestFloatHistogramEquals(t *testing.T) {
+	// mark this test case as being run in parallel
+	t.Parallel()
 	h1 := FloatHistogram{
 		Schema:          3,
 		Count:           2.2,
@@ -3454,7 +3511,11 @@ func TestFloatHistogramSize(t *testing.T) {
 	}
 
 	for _, c := range cases {
+		// capture the range variable to avoid issues with concurrency
+		c := c
 		t.Run(c.name, func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			require.Equal(t, c.expected, c.fh.Size())
 		})
 	}
@@ -3505,7 +3566,11 @@ func TestFloatHistogramString(t *testing.T) {
 	}
 
 	for _, c := range cases {
+		// capture the range variable to avoid issues with concurrency
+		c := c
 		t.Run(c.name, func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			require.NoError(t, c.fh.Validate())
 			require.Equal(t, c.expected, c.fh.String())
 		})
@@ -3597,10 +3662,16 @@ func TestFloatHistogramReduceResolution(t *testing.T) {
 		},
 	}
 
-	for _, tc := range tcs {
-		target := tc.origin.ReduceResolution(tc.target.Schema)
-		require.Equal(t, tc.target, target)
-		// Check that receiver histogram was mutated:
-		require.Equal(t, tc.target, tc.origin)
+	for testName, tc := range tcs {
+		// capture the range variable to avoid issues with concurrency
+		tc := tc
+		t.Run(testName, func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
+			target := tc.origin.ReduceResolution(tc.target.Schema)
+			require.Equal(t, tc.target, target)
+			// Check that receiver histogram was mutated:
+			require.Equal(t, tc.target, tc.origin)
+		})
 	}
 }