--- a/model/histogram/histogram_test.go
+++ b/model/histogram/histogram_test.go
@@ -87,7 +87,11 @@ func TestHistogramString(t *testing.T) {
 	}
 
 	for i, c := range cases {
+		// capture the range variable to avoid issues with concurrency
+		c := c
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			actualString := c.histogram.String()
 			require.Equal(t, c.expectedString, actualString)
 		})
@@ -246,7 +250,11 @@ func TestCumulativeBucketIterator(t *testing.T) {
 	}
 
 	for i, c := range cases {
+		// capture the range variable to avoid issues with concurrency
+		c := c
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			it := c.histogram.CumulativeBucketIterator()
 			actualBuckets := make([]Bucket[uint64], 0, len(c.expectedBuckets))
 			for it.Next() {
@@ -462,7 +470,11 @@ func TestRegularBucketIterator(t *testing.T) {
 	}
 
 	for i, c := range cases {
+		// capture the range variable to avoid issues with concurrency
+		c := c
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			it := c.histogram.PositiveBucketIterator()
 			actualPositiveBuckets := make([]Bucket[uint64], 0, len(c.expectedPositiveBuckets))
 			for it.Next() {
@@ -545,7 +557,11 @@ func TestHistogramToFloat(t *testing.T) {
 	}
 
 	for _, c := range cases {
+		// capture the range variable to avoid issues with concurrency
+		c := c
 		t.Run(c.name, func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			fh := h.ToFloat(c.fh)
 			require.Equal(t, h.String(), fh.String())
 		})
@@ -612,7 +628,11 @@ func TestCustomBucketsHistogramToFloat(t *testing.T) {
 
 	require.NoError(t, h.Validate())
 	for _, c := range cases {
+		// capture the range variable to avoid issues with concurrency
+		c := c
 		t.Run(c.name, func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			hStr := h.String()
 			fh := h.ToFloat(c.fh)
 			require.NoError(t, fh.Validate())
@@ -624,6 +644,8 @@ func TestCustomBucketsHistogramToFloat(t *testing.T) {
 
 // TestHistogramEquals tests both Histogram and FloatHistogram.
 func TestHistogramEquals(t *testing.T) {
+	// mark this test case as being run in parallel
+	t.Parallel()
 	h1 := Histogram{
 		Schema:        3,
 		Count:         62,
@@ -867,7 +889,11 @@ func TestHistogramCopy(t *testing.T) {
 	}
 
 	for _, tcase := range cases {
+		// capture the range variable to avoid issues with concurrency
+		tcase := tcase
 		t.Run(tcase.name, func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			hCopy := tcase.orig.Copy()
 
 			// Modify a primitive value in the original histogram.
@@ -932,7 +958,11 @@ func TestHistogramCopyTo(t *testing.T) {
 	}
 
 	for _, tcase := range cases {
+		// capture the range variable to avoid issues with concurrency
+		tcase := tcase
 		t.Run(tcase.name, func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			hCopy := &Histogram{}
 			tcase.orig.CopyTo(hCopy)
 
@@ -945,6 +975,7 @@ func TestHistogramCopyTo(t *testing.T) {
 }
 
 func assertDeepCopyHSpans(t *testing.T, orig, hCopy, expected *Histogram) {
+	t.Helper()
 	// Do an in-place expansion of an original spans slice.
 	orig.PositiveSpans = expandSpans(orig.PositiveSpans)
 	orig.PositiveSpans[len(orig.PositiveSpans)-1] = Span{1, 2}
@@ -1293,7 +1324,11 @@ func TestHistogramCompact(t *testing.T) {
 	}
 
 	for _, c := range cases {
+		// capture the range variable to avoid issues with concurrency
+		c := c
 		t.Run(c.name, func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			require.Equal(t, c.expected, c.in.Compact(c.maxEmptyBuckets))
 			// Compact has happened in-place, too.
 			require.Equal(t, c.expected, c.in)
@@ -1568,7 +1603,11 @@ func TestHistogramValidation(t *testing.T) {
 	}
 
 	for testName, tc := range tests {
+		// capture the range variable to avoid issues with concurrency
+		tc := tc
 		t.Run(testName, func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			if err := tc.h.Validate(); tc.errMsg != "" {
 				require.EqualError(t, err, tc.errMsg)
 			} else {
@@ -1633,10 +1672,14 @@ func TestHistogramReduceResolution(t *testing.T) {
 		},
 	}
 
-	for _, tc := range tcs {
-		target := tc.origin.ReduceResolution(tc.target.Schema)
-		require.Equal(t, tc.target, target)
-		// Check that receiver histogram was mutated:
-		require.Equal(t, tc.target, tc.origin)
+	for caseName, tc := range tcs {
+		// capture the range variable to avoid issues with concurrency
+		tc := tc
+		t.Run(caseName, func(t *testing.T) {
+			target := tc.origin.ReduceResolution(tc.target.Schema)
+			require.Equal(t, tc.target, target)
+			// Check that receiver histogram was mutated:
+			require.Equal(t, tc.target, tc.origin)
+		})
 	}
 }