--- a/model/labels/matcher_test.go
+++ b/model/labels/matcher_test.go
@@ -22,6 +22,7 @@ import (
 )
 
 func mustNewMatcher(t *testing.T, mType MatchType, value string) *Matcher {
+	t.Helper()
 	m, err := NewMatcher(mType, "test_label_name", value)
 	require.NoError(t, err)
 	return m
@@ -100,8 +101,14 @@ func TestMatcher(t *testing.T) {
 		},
 	}
 
-	for _, test := range tests {
-		require.Equal(t, test.matcher.Matches(test.value), test.match)
+	for i, test := range tests {
+		// capture the range variable to avoid issues with concurrency
+		test := test
+		t.Run(fmt.Sprintf("%d: %s", i, test.matcher), func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
+			require.Equal(t, test.matcher.Matches(test.value), test.match)
+		})
 	}
 }
 
@@ -128,10 +135,16 @@ func TestInverse(t *testing.T) {
 		},
 	}
 
-	for _, test := range tests {
-		result, err := test.matcher.Inverse()
-		require.NoError(t, err)
-		require.Equal(t, test.expected.Type, result.Type)
+	for i, test := range tests {
+		// capture the range variable to avoid issues with concurrency
+		test := test
+		t.Run(fmt.Sprintf("%d: %s", i, test.matcher), func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
+			result, err := test.matcher.Inverse()
+			require.NoError(t, err)
+			require.Equal(t, test.expected.Type, result.Type)
+		})
 	}
 }
 
@@ -181,7 +194,11 @@ func TestPrefix(t *testing.T) {
 			prefix:  "",
 		},
 	} {
+		// capture the range variable to avoid issues with concurrency
+		tc := tc
 		t.Run(fmt.Sprintf("%d: %s", i, tc.matcher), func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			require.Equal(t, tc.prefix, tc.matcher.Prefix())
 		})
 	}
@@ -205,7 +222,11 @@ func TestIsRegexOptimized(t *testing.T) {
 			isRegexOptimized: true,
 		},
 	} {
+		// capture the range variable to avoid issues with concurrency
+		tc := tc
 		t.Run(fmt.Sprintf("%d: %s", i, tc.matcher), func(t *testing.T) {
+			// mark this test case as being run in parallel
+			t.Parallel()
 			require.Equal(t, tc.isRegexOptimized, tc.matcher.IsRegexOptimized())
 		})
 	}