… model/labels/labels_test.go
… model/labels/regexp_test.go
… model/labels/sharding_test.go
… model/relabel/relabel_test.go
… model/rulefmt/rulefmt_test.go
… model/textparse/interface_test.go
… model/textparse/nhcbparse_test.go
… model/textparse/openmetricsparse_test.go
… model/textparse/promparse_test.go
… model/textparse/protobufparse_test.go