I have added `t.Parallel()` to most tests in the models package as part of the checklist outlined in #15185.

Below is a comparison of execution times before and after these changes.

Before:

```bash
$ go test --tags=dedupelabels -race --count=1 ./model/...
?       github.com/prometheus/prometheus/model/exemplar [no test files]
?       github.com/prometheus/prometheus/model/metadata [no test files]
ok      github.com/prometheus/prometheus/model/histogram        1.095s
?       github.com/prometheus/prometheus/model/timestamp        [no test files]
?       github.com/prometheus/prometheus/model/value    [no test files]
ok      github.com/prometheus/prometheus/model/labels   3.038s
ok      github.com/prometheus/prometheus/model/relabel  1.086s
ok      github.com/prometheus/prometheus/model/rulefmt  1.091s
ok      github.com/prometheus/prometheus/model/textparse        1.293s
```

After:

```bash
$ go test --tags=dedupelabels -race --count=1 ./model/...
?       github.com/prometheus/prometheus/model/exemplar [no test files]
?       github.com/prometheus/prometheus/model/metadata [no test files]
?       github.com/prometheus/prometheus/model/timestamp        [no test files]
?       github.com/prometheus/prometheus/model/value    [no test files]
ok      github.com/prometheus/prometheus/model/histogram        1.059s
ok      github.com/prometheus/prometheus/model/labels   2.792s
ok      github.com/prometheus/prometheus/model/relabel  1.046s
ok      github.com/prometheus/prometheus/model/rulefmt  1.040s
ok      github.com/prometheus/prometheus/model/textparse        1.270s
```

The performance improvements are not as significant as in related PRs, but there is still a noticeable gain. Some tests cannot run in parallel due to race conditions caused by the use of global variables. Additionally, I left Benchmark tests in certain subpackages untouched. ~Interestingly, the execution time for the textparse package increased slightly, though the exact reason is unclear.~

Edit: I realized that some of the tests require a build label, which I had initially overlooked. I've updated the test results accordingly, and now all the subpackages show a gain, even if it's a slight one.