@jvalecillos 2024-11-19
I rebased my changes from the latest `main` but now a test in the `tracing` package (not related with my changes) is failing. I'm looking into it.