@jvalecillos 2024-11-19
As I suspected, using git commit --amend --no-edit and later pushing to trigger a new build resulted in all the tests passing. This suggests either that the CI doesn't always execute tests consistently or that some tests aren't idempotent. Personally, I believe it's the latter.