@jvalecillos 2025-02-11
Rebased from `main` branch to solve conflicts on updated tests