@jvalecillos 2025-02-28
> I think this is a lot of added bulk for benefits that seem small.
> 
> For example, all the tests in `model/labels` run in 120ms total. We don't need to speed that up at the cost of tens of lines of code.
> 
> Can you pick one or two packages where you would argue this is worth doing?

I understand that priorities can change, but after being encouraged to work on this issue and making multiple revisions based on the requested feedback, it's disappointing to now hear that the benefits are considered too small. Given that other project members initially saw value in this work in #15185, I would appreciate some clarity on what changed. I don’t mean to be rude, but I’d like to understand the reasoning behind this decision, especially after investing so much time into it.