@machine424 2025-02-28
José,  
Thanks for working on this!  

Maybe the issue https://github.com/prometheus/prometheus/issues/15185 wasn’t clear. If the task were simply to add `t.Parallel()` here and there, I could have done that myself without opening a ticket.

The task requires more than that. I created the issue as an entry point for potential new contributors. It helps them familiarize themselves with the tests, the CI, the codebase...; have their first contribution to Prometheus.

In the issue, I listed all the packages **to check** for potential test speed improvements. Below the list, I included some criteria. One of the them is:

`Let’s say it’s worthwhile when we can save at least one second per package; anything less might just be fluctuation.`
(maybe we should look for more that 1s)
  
As you can see, the `config/` package was crossed off the list because it turned out that parallelization wasn’t beneficial. 

As Brian suggested, let’s cross that one off and pick another package, if you're still interested of course.

Time spent learning is never wasted ;)